Replace mounts with more secure volume syntax
Also, make them read only to disallow write access in a proper way. Before, @IceflowRE relied on the Docker user to not have permissions to those directories. But in theory, both the outside user ID and the Docker container user ID might be the same, which then bypasses the "protection". The read-only volumes are anyway more secure (volume > mount).